#4: strip index to base; wrap in layout; move pages to temp add base dark style fix error in layout.tsx, .css
